% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{dispersionDGE}
\alias{dispersionDGE}
\title{Plot Dispersion}
\usage{
dispersionDGE(data, ...)
}
\arguments{
\item{data}{data.table; the output of summariseDGE}
}
\value{
A ggplot2 object
}
\description{
\code{dispersionDGE} Scatter plot of dispersion by gene mean
}
\examples{
# see vignette

}
