% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{dropoutsDGE}
\alias{dropoutsDGE}
\title{Plot Dropout Rate}
\usage{
dropoutsDGE(data, K = 1.056)
}
\arguments{
\item{data}{data.table; the output of summariseDGE}

\item{K}{numeric; Km parameter of the michaelis menten equation shown in red in the output}
}
\value{
A ggplot2 object
}
\description{
\code{dropoutsDGE} Plot dropout rate by gene mean
}
\examples{
# see vignette

}
