% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_parameters}
\alias{fit_parameters}
\title{Fit parameters to a dataset}
\usage{
fit_parameters(data, plot = TRUE, method = "mge", ...)
}
\arguments{
\item{data}{matrix; gene by cell dge matrix of counts}

\item{plot}{logical; if TRUE the real vs theoretical distribution plots qqplot etc. are printed}
}
\value{
An object of class dropsim_parameters containing fitted parameters for the mean and library size lognormal distributions
}
\description{
\code{fit_parameters} find parameters from data to simulate new dataset
}
\examples{
# new_parameters <- fit_parameters(your_data_matrix)

}
