% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{histogramDGE}
\alias{histogramDGE}
\title{Plot Mean Histogram}
\usage{
histogramDGE(data)
}
\arguments{
\item{data}{data.table; the output of summariseDGE}
}
\value{
A ggplot2 object
}
\description{
\code{histogramDGE} Plot a histogram of the gene means
}
\examples{
# see vignette

}
