% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{mean_varianceDGE}
\alias{mean_varianceDGE}
\title{Plot Mean Variance Relationship}
\usage{
mean_varianceDGE(data)
}
\arguments{
\item{data}{data.table; the output of summariseDGE}
}
\value{
A ggplot2 object
}
\description{
\code{histogramDGE} Scatter plot of Variance by Mean
}
\examples{
# see vignette

}
