% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{normalised_varianceDGE}
\alias{normalised_varianceDGE}
\title{Plot Normalised Variance}
\usage{
normalised_varianceDGE(data)
}
\arguments{
\item{data}{data.table; the output of summariseDGE}
}
\value{
A ggplot2 object
}
\description{
\code{histogramDGE} Scatter plot of Variance/Mean by Mean
}
\examples{
# see vignette

}
