% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{normaliseDGE}
\alias{normaliseDGE}
\title{Normalise Digital Gene Expression Matrix}
\usage{
normaliseDGE(dge, verbose = FALSE, center = TRUE, scale = TRUE,
  threshold = 5, min_library_size = 1)
}
\arguments{
\item{dge}{matrix; a digital gene expression matrix containing count data}

\item{verbose}{logical; if TRUE the function will print diagnostic graphs along the way}

\item{center}{logical; if true genes are centered to mean of 0}

\item{scale}{logical; if TRUE genes are scaled to unit variance
(If sparse=FALSE, setting dge=FALSE can save time on larger datasets.)}

\item{threshold}{integer; any values larger than this will be rounded down}
}
\value{
A matrix of read counts
}
\description{
\code{normaliseDGE} simulate digital gene expression matrix containing count data from given parameters
}
\examples{
# normalised_counts <- normalise_dge(dge)
}
