% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_summaryDGE}
\alias{plot_summaryDGE}
\title{Plot Summary Grid}
\usage{
plot_summaryDGE(data)
}
\arguments{
\item{data}{data.table; the output of summariseDGE}
}
\value{
A ggplot2 object
}
\description{
\code{plot_summaryDGE} Plot multiple gene level summaries
}
\examples{
# see vignette

}
