% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulateDGE}
\alias{simulateDGE}
\title{Simulate Single Cell RNAseq count data}
\usage{
simulateDGE(parameters, sparse = TRUE, cell_prefix = "cell", dge = TRUE)
}
\arguments{
\item{parameters}{object of class sinsynthr containing parameters}

\item{sparse}{logical; if true the counts are returned as a sparse matrix}

\item{cell_prefix}{character; the default group name for cells}

\item{dge}{logical; if false counts a returned as a matrix cells by genes rather than genes by cells
(If sparse=FALSE, setting dge=FALSE can save time on larger datasets.)}
}
\value{
A matrix of read counts
}
\description{
\code{simulateDGE} simulate digital gene expression matrix containing count data from given parameters
}
\examples{
# new_parameters <- sinsynthr_parameters()
# dge <- simulate(new_parameters)
}
