% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\docType{class}
\name{sinsynthr_parameters-class}
\alias{sinsynthr_parameters-class}
\alias{sinsynthr_parameters}
\title{An S4 class to hold simulation parameters.}
\description{
An S4 class to hold simulation parameters.
}
\section{Slots}{

\describe{
\item{\code{n_genes}}{integer; number of genes to simulate}

\item{\code{n_cells}}{integer; number of cells to simulate}

\item{\code{gene_meanlog}}{numeric; meanlog of lognormal distribution to simulate means from, see ?rlnorm for more details}

\item{\code{gene_sdlog}}{numeric; sdlog of lognormal distribution to simulate means from, see ?rlnorm for more details}

\item{\code{library_meanlog}}{numeric; meanlog of lognormal distribution to simulate library sizes from, see ?rlnorm for more details}

\item{\code{library_sdlog}}{numeric; sdlog of lognormal distribution to simulate library sizes from, see ?rlnorm for more details}

\item{\code{groups}}{data.table with variables 'scale', 'cells' and 'names'
scale is a numeric vector of the scale parameter for the logistic distribution to simulate the fold changes from
cells is a list of integer vector giving the indexes of cells each differential expression should be applied to, cells can be a member of more than one group
names is a character vector of cell group names}
}}

