% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{summariseDGE}
\alias{summariseDGE}
\title{Calculate gene wise summaries}
\usage{
summariseDGE(data, name = NA, rm.0 = TRUE)
}
\arguments{
\item{data}{matrix; A digital gene expression matrix, e.g. output from simulateDGE()}

\item{name}{character; A descriptive name for the dataset}

\item{rm.0}{logical; if TRUE genes with a mean of 0 will be removed}
}
\value{
A data.table of summary statistics of each gene
}
\description{
\code{summariseDGE} Calculate gene wise summaries from a Digital Gene Expression matrix
}
